helpers refactor
